requirements.txt
asgiref==3.2.7
autopep8==1.5.3
Django==3.0.7
django-cors-headers==3.4.0
django-username-email==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
